% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\docType{data}
\name{Copper}
\alias{Copper}
\title{Mining data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 442 rows and 7 columns.}
\usage{
data( 'Copper', packages = 'KRIG' )
}
\description{
Mining data
}
\examples{
data( 'Copper', packages = 'KRIG' )
}
\references{
Clark
}
\keyword{kriging}
