% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exp_kernel}
\alias{exp_kernel}
\title{Exponential kernel.}
\usage{
exp_kernel(h, sigma = 1, theta = 1)
}
\arguments{
\item{h}{distance variable.}

\item{sigma}{amplitude parameter.}

\item{theta}{smoothness parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernel.
}
\examples{
h<-0.3
sigma<-1.0
theta<-10
exp_kernel( h, sigma, theta )
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
