% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Gaussian kernel.}
\usage{
gaussian_kernel(h, sigma = 1, theta = 1)
}
\arguments{
\item{h}{distance variable.}

\item{sigma}{amplitude parameter.}

\item{theta}{smoothness parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernel.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
