% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Kov}
\alias{Kov}
\title{Spatial covariance matrix.}
\usage{
Kov(X, Y, Kern, symmetric = FALSE)
}
\arguments{
\item{X}{First set of spatial points.}

\item{Y}{Second set of spatial points.}

\item{Kern}{Kernel function.}

\item{symmetric}{If result of computation will be a square matrix, the time computation can
be improved setting this parameter to TRUE, the default is FALSE.}
}
\value{
The spatial covariance matrix.
}
\description{
To compute a kriging, it is necessary the spatial covariance matrix. The spatial 
  covariance could computed between to sets of points X and Y with different dimension and the 
  result it is not necessarily a square matrix.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
