% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Krigidx}
\alias{Krigidx}
\title{Combinatorial variance computation.}
\usage{
Krigidx(KF, comb, X, Gamma)
}
\arguments{
\item{KF}{values of the kernel integral evaluations.}

\item{comb}{Combination.}

\item{X}{Points in the grid.}

\item{Gamma}{Cube with integral results.}
}
\value{
Real value of sensitivity.
}
\description{
For a given combination this function computes the associated variance for the
variable enumerated by the combination values.
}
\references{
\insertRef{Kanova:2013}{KRIG}
}
\seealso{
For a complete application you can check the documentation of \code{\link{Krigvar}}.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
