% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_kernel}
\alias{linear_kernel}
\title{Linear kernel}
\usage{
linear_kernel(x, y, alpha)
}
\arguments{
\item{x}{first column vector.}

\item{y}{second column vector.}

\item{alpha}{amplitude parameter.}
}
\value{
Real value.
}
\description{
Anisotropic kernel defined by the scalar product.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
