% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_integrate_kernel}
\alias{list_integrate_kernel}
\title{Integrals of a list of kernels.}
\usage{
list_integrate_kernel(Kernels, X)
}
\arguments{
\item{Kernels}{data.frame of kernels composed by four columns.
\enumerate{
  \item Kernel name.
  \item Inferior limit for integral.
  \item Superior limit for integral.
  \item Number of steps for discretization of integrals.
}}

\item{X}{matrix containing in each row the coordinate where the one coordinate integrals will 
be evaluated.}
}
\value{
List with one coordinate integrals and complete kernel integrals.
}
\description{
This function is part of the routines employed in the sensitivity analysis, 
computes vector of integrals and complete integrals of kernels specified in the data frame
Kernels.
}
\seealso{
For a complete application you can check the documentation of \code{\link{Kanova}}.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
