% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matern_kernel}
\alias{matern_kernel}
\title{Matérn kernel.}
\usage{
matern_kernel(h, v = 2, sigma = 1, theta = 1)
}
\arguments{
\item{h}{distance variable.}

\item{v}{power parameter.}

\item{sigma}{amplitude parameter.}

\item{theta}{smootheness parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernal.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
