% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multilog_kernel}
\alias{multilog_kernel}
\title{Multilog kernel.}
\usage{
multilog_kernel(h, R = 1)
}
\arguments{
\item{h}{distance parameter.}

\item{R}{displacement parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernel.
}
\examples{
h<-4
R<-2.3
multilog_kernel( h, R )
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
