% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polynomial_kernel}
\alias{polynomial_kernel}
\title{Polynomial kernel}
\usage{
polynomial_kernel(x, y, alpha, beta, n)
}
\arguments{
\item{x}{first column vector.}

\item{y}{second column vector.}

\item{alpha}{amplitude parameter.}

\item{beta}{displacement parameter.}

\item{n}{power parameter.}
}
\value{
Real value.
}
\description{
Anisotropic kernel defined like a polynomial in the scalar product.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
N<-10
x<-matrix( runif( N ), N, 1 )
y<-matrix( runif( N ), N, 1 )
alpha<-0.5
beta<-2
n<-3
polynomial_kernel( x, y, alpha, beta, n )
}
