% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHAnova}
\alias{RKHAnova}
\title{Integrals of kernels}
\usage{
RKHAnova(Kernels, Integral, X)
}
\arguments{
\item{Kernels}{List of kernels}
}
\value{
List with vector integrals and complete kernel integrals
}
\description{
Compute integrals of kernels
}
\author{
Pedro Guarderas
}
