% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHCov}
\alias{RKHCov}
\title{Covariance kernel matrix}
\usage{
RKHCov(X, Y, k, symmetric = FALSE)
}
\arguments{
\item{k}{Kernel}

\item{symmetric}{check if matrix will be symmetric}
}
\value{
Covariance matrix
}
\description{
Computes covariance kernel matrix
}
\author{
Pedro Guarderas
}
