% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHEstimate}
\alias{RKHEstimate}
\title{Gaussian regression}
\usage{
RKHEstimate(Z, X, Y, K, k, S, type = 1L, cinv = 0L)
}
\arguments{
\item{Z}{Observed spatial values.}

\item{X}{Spatial points, ubication where the Z value was observed.}

\item{Y}{Spatial points where it is requested to predict Z.}

\item{type}{Type of Gaussian regression, 0 = simple interpolation, 1 = simple regression, 
2 = ordinary interpolation, 3 = ordinary regression, 4 = universal interpolation, 
5 = universal regression.}

\item{cinv}{Form to compute the inverse.}

\item{Kern}{Kernel functions.}
}
\value{
List containing the different important matrices employed for the Gaussian regression.
\item{W}{Predicted new values for Z}
\item{K}{Covariance matrix for X}
\item{k}{Covariance matrix between Y and X}
\item{J}{Inverse of the covariance matrix}
}
\description{
Computes Gaussian regression with a given covariance kernel
}
\author{
Pedro Guarderas
}
