% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHGaussProcess}
\alias{RKHGaussProcess}
\title{Gaussian regression}
\usage{
RKHGaussProcess(Z, X, Y, Kern)
}
\arguments{
\item{Y}{}

\item{Kern}{kernel}
}
\value{
List
}
\description{
Computes Gaussian regression with a given covariance kernel
}
\author{
Pedro Guarderas
}

