% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHIntegrateKern}
\alias{RKHIntegrateKern}
\title{Vector integral}
\usage{
RKHIntegrateKern(Kern, x, a, b, n)
}
\value{
Vector with integrals
}
\description{
Computes covariance kernel matrix
}
\author{
Pedro Guarderas
}
