\name{RKHSENS}
\alias{RKHSENS}
\alias{RKHSENS}
\docType{package}
\title{
RKHSENS
}
\description{
Package for sensitivity analysis
}
\details{
\tabular{ll}{
Package: \tab RKHSENS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-18\cr
License: \tab GPL (>= 2)\cr
}
Kriging modeling simple and ordinary. SSANOVA analysis, Sobol indices computation.
}
\author{Julio Andrade, Daniel Lagos, Andrés Lopez, Pedro Guarderas, Paul Recalde, Edison Salazar
Maintainer: Pedro Guarderas <ajusworkopensource@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
