% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHSobolVar}
\alias{RKHSobolVar}
\title{Sensitiviy analysis}
\usage{
RKHSobolVar(Gamma)
}
\arguments{
\item{Gamma}{Cube with integral results.}
}
\value{
Real value of sensitivity.
}
\description{
Computation of variance
}
\author{
Pedro Guarderas
}
