% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sens_idx}
\alias{sens_idx}
\title{Sensitiviy analysis}
\usage{
sens_idx(KF, comb, X, Gamma)
}
\arguments{
\item{KF}{}

\item{comb}{Combination.}

\item{X}{Points grid.}

\item{Gamma}{Cube with integral results.}
}
\value{
Real value of sensitivity.
}
\description{
Computation of Sobol index
}
\author{
Pedro Guarderas
}
