% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sens_var}
\alias{sens_var}
\title{Combinatorial variance computation.}
\usage{
sens_var(KF, Gamma)
}
\arguments{
\item{KF}{values of the kernel integral evaluations.}

\item{Gamma}{Cube with integral results.}
}
\value{
Real value of sensitivity.
}
\description{
Computation of variance
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
