% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{square_kernel}
\alias{square_kernel}
\title{Square kernel.}
\usage{
square_kernel(h, alpha = 1)
}
\arguments{
\item{h}{distance variable.}

\item{alpha}{amplitude parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernel given by the square distance.
}
\examples{
h<-4
alpha<-2
square_kernel( h, alpha )
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
