% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{variogram}
\alias{variogram}
\title{Computes the variogram.}
\usage{
variogram(Z, X, d)
}
\arguments{
\item{Z}{Vector of observations.}

\item{X}{Points matrix.}

\item{d}{Distance function.}
}
\value{
Variogram vector.
}
\description{
This useful function is commonly employed in the study of isotropic stationary 
spatial processes.
}
\examples{
library( KRIG )
vignette( topic = 'copper_mining_2d', package = 'KRIG' )
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
