% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_integrate_kernel}
\alias{vector_integrate_kernel}
\title{Integrals of kernels}
\usage{
vector_integrate_kernel(Kernels, X)
}
\arguments{
\item{Kernels}{List of kernels}
}
\value{
List with vector integrals and complete kernel integrals
}
\description{
Compute integrals of kernels
}
\author{
Pedro Guarderas
}
