% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{complete_integrate_kernel}
\alias{complete_integrate_kernel}
\title{Complete kernel integral}
\usage{
complete_integrate_kernel(Kern, a, b, n)
}
\value{
Real
}
\description{
Complete kernel integral
}
\author{
Pedro Guarderas
}
