% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integrate_kernel}
\alias{integrate_kernel}
\title{Vector integral}
\usage{
integrate_kernel(Kern, x, a, b, n)
}
\value{
Vector with integrals
}
\description{
Computes covariance kernel matrix
}
\author{
Pedro Guarderas
}
