% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Kanova}
\alias{Kanova}
\title{KANOVA, kernel anova under RKHS approximations.}
\usage{
Kanova(Kernels, Integral, X)
}
\arguments{
\item{Kernels}{data.frame of kernels composed by four columns.
\enumerate{
  \item Kernel name.
  \item Inferior limit for integral.
  \item Superior limit for integral.
  \item Number of steps for discretization of integrals.
}}

\item{Integral}{A list containing the results of kernel integrals of the functions
\code{\link{vector_integrate_kernel}}.}

\item{X}{matrix containing in each row the coordinate where the one coordinate integrals will 
be evaluated.}
}
\value{
List with containing the Gamma 3D array where the different combination variance are
stocked and the total matrix variance named Kanova.
}
\description{
Under an approximation to the sensitivity analysis based in variance computation
the different indexes of combinatorial sensitivity values can be computed employing the
values of kernel integrals.
}
\examples{
library( KRIG )
options( stringsAsFactors = FALSE )
 
kernel_1<-function( x, y ) exp( -0.5*(x-y)^2)
kernel_2<-function( x, y ) exp( -0.7*(x-y)^2)
kernel_3<-function( x, y ) exp( -0.1*(x-y)^2)

Kernels<-data.frame( kernel = c( 'kernel_1', 'kernel_2', 'kernel_3' ), 
                     min = c( -1, -1, -5 ), 
                     max = c( 1, 1, 5 ),
                     n = c( 500, 500, 500 ) )
                     
n<-20
X<-matrix( c( seq( -1, 1, length.out = n ), 
              seq( -1, 1, length.out = n ),
              seq( -5, 5, length.out = n ) ), n, 3 )
              
KI<-list_integrate_kernel( Kernels, X )
GK<-Kanova( Kernels, KI, X )
    
f<-function( x ) x[1] + 30 * x[2] + 60 * x[3]
Func<-apply( X, 1, FUN = f )
    
KF<-solve( GK$Kanova + diag( 1e-8, n, n ), Func )
}
\references{
\insertRef{Kanova:2013}{KRIG}
\insertRef{Aronszajn:ThRK}{KRIG}
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
