% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matern_kernel}
\alias{matern_kernel}
\title{Matérn kernel.}
\usage{
matern_kernel(h, v = 2, sigma = 1, theta = 1)
}
\arguments{
\item{h}{distance variable.}

\item{v}{power parameter.}

\item{sigma}{amplitude parameter.}

\item{theta}{smoothness parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernel.
}
\examples{
h<-4.0
v<-2.0
sigma<-2.0
theta<-100.0
matern_kernel( h, v, sigma, theta )
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
