% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nat_cubic_spline_kernel}
\alias{nat_cubic_spline_kernel}
\title{Natural cubic spline kernel.}
\usage{
nat_cubic_spline_kernel(h, R = 1)
}
\arguments{
\item{h}{distance variable.}

\item{R}{displacement parameter.}
}
\value{
Real value.
}
\description{
Isotropic kernel.
}
\examples{
h<-4.0
R<-2.3
nat_cubic_spline_kernel( h, R )
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
