% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHCov}
\alias{RKHCov}
\title{Spatial covariance matrix.}
\usage{
RKHCov(X, Y, Kern, symmetric = FALSE)
}
\arguments{
\item{X}{First set of spatial points.}

\item{Y}{Second set of spatial points.}

\item{Kern}{Kernel function.}

\item{symmetric}{If result of computation will be a square matrix, the time computation can
be improved setting this parameter to TRUE, the default is FALSE.}
}
\value{
The spatial covariance matrix.
}
\description{
To compute a Gaussian regression or interpolation is necessary use Kernel, and
  it is requeried to construct the spatial convariance matrix. The spatial covariance could
  be estimated between to sets of points X and Y and the result is not necessarily a square
  matrix.
}
\author{
Pedro Guarderas
}
