% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHEstimate}
\alias{RKHEstimate}
\title{Gaussian regression}
\usage{
RKHEstimate(Z, X, Y, Kern, type = 1L)
}
\arguments{
\item{Y}{}

\item{Kern}{kernel}
}
\value{
List
}
\description{
Computes Gaussian regression with a given covariance kernel
}
\author{
Pedro Guarderas
}
