% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RKHWeightPowDist}
\alias{RKHWeightPowDist}
\title{Generic function to compute distances}
\usage{
RKHWeightPowDist(x, y, w, p)
}
\arguments{
\item{w}{weights}

\item{p}{power terms}
}
\value{
Numeric value with the distance
}
\description{
This function computes a weighted distance between vectors x and y
}
\author{
Pedro Guarderas
}

