% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{variogram}
\alias{variogram}
\title{Computes the variogram.}
\usage{
variogram(Z, X, d, delta)
}
\arguments{
\item{Z}{Vector of observations.}

\item{X}{Points matrix.}

\item{d}{Distance function.}

\item{delta}{Search distance ratio.}
}
\value{
Variogram vector.
}
\description{
This useful function is commonly employed in the study of isotropic stationary 
spatial processes.
}
\author{
Pedro Guarderas \email{pedro.felipe.guarderas@gmail.com}.
}
